Refactor initial client base code added folders for each widget added test folders for each corresponding widget
